create guestapp and move guset lectures to here
