+++ speakers/guestapp/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from .views import LecturesAPIView
+
+urlpatterns = [
+    path('lecture/', LecturesAPIView.as_view()),
+]
