+++ speakers/speakers/urls.py
@@ -5,7 +5,6 @@ from django.urls import path
 from django.urls import include
 
 from .utils.yasg import urlpatterns as swagger_urls
-from workroomsapp.views.guest_views import *
 from .views import index
 
 urlpatterns = [
@@ -19,7 +18,7 @@ urlpatterns = [
 
     path('api/auth/', include('authapp.urls')),
     path('api/workrooms/', include('workroomsapp.urls')),
-    path('api/lecture/', LecturesAPIView.as_view()),
+    path('api/guest/', include('guestapp.urls')),
 ]
 
 if settings.DEBUG:
